elt_project2/requirements.txt
pandas
psycopg2-binary
